Extract ParsedVariable type to new file.
